Move ChartToggle to components folder
