+++ src/components/ChartToggle/index.tsx
@@ -1,0 +1,1 @@
+export { ChartToggle } from './ChartToggle'
